# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects": {
      "get": {
        "operationId": "ProjectsController_findAll",
        "summary": "List all projects",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "example": "villa",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "planning",
              "type": "string"
            }
          },
          {
            "name": "projectManagerId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "uuid-user",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all projects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "operationId": "ProjectsController_create",
        "summary": "Create new project",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findOne",
        "summary": "Get project details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return project details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "operationId": "ProjectsController_updateProject",
        "summary": "Update project details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{id}/tasks": {
      "get": {
        "operationId": "ProjectsController_findTasks",
        "summary": "Get project tasks",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return list of project tasks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{id}/activity": {
      "get": {
        "operationId": "ProjectsController_findActivity",
        "summary": "Get project activity feed",
        "description": "Returns a chronological log of all changes and actions taken on this project.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activity feed retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/api/rbac/roles/{id}/permissions": {
      "post": {
        "operationId": "RbacController_updateRolePermissions",
        "summary": "Update role permissions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRolePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "RBAC"
        ]
      }
    },
...
    "/api/meta/entities/{entityKey}": {
      "get": {
        "operationId": "RuntimeController_getEntity",
        "summary": "Load runtime entity metadata",
        "parameters": [
          {
            "name": "entityKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Runtime"
        ]
      }
    },
    "/api/data/{entityKey}/list": {
      "post": {
        "operationId": "RuntimeController_listRecords",
        "summary": "Generic entity list",
        "parameters": [
          {
            "name": "entityKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Runtime"
        ]
      }
    },
    "/api/data/{entityKey}/get": {
      "post": {
        "operationId": "RuntimeController_getRecord",
        "summary": "Generic entity get",
        "parameters": [
          {
            "name": "entityKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Runtime"
        ]
      }
    },
    "/api/data/{entityKey}/create": {
      "post": {
        "operationId": "RuntimeController_createRecord",
        "summary": "Generic entity create",
        "parameters": [
          {
            "name": "entityKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Runtime"
        ]
      }
    },
    "/api/data/{entityKey}/update": {
      "post": {
        "operationId": "RuntimeController_updateRecord",
        "summary": "Generic entity update",
        "parameters": [
          {
            "name": "entityKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Runtime"
        ]
      }
    },
    "/api/data/{entityKey}/delete": {
      "post": {
        "operationId": "RuntimeController_deleteRecord",
        "summary": "Generic entity delete",
        "parameters": [
          {
            "name": "entityKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Runtime"
        ]
      }
    },
    "/api/data/{entityKey}/bulk": {
      "post": {
        "operationId": "RuntimeController_bulkRecord",
        "summary": "Generic entity bulk operation",
        "parameters": [
          {
            "name": "entityKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Runtime"
        ]
      }
    },
    "/api/action/{actionKey}": {
      "post": {
        "operationId": "RuntimeController_executeAction",
        "summary": "Execute a generic CRM action",
        "parameters": [
          {
            "name": "actionKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Runtime"
        ]
      }
    },
...
  },
  "info": {
    "title": "CRM Backend API",
    "description": "The core API for CRM Lead Management and Project Coordination.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "crm_session",
        "description": "Session cookie (HttpOnly)"
      }
    },
    "schemas": {
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "rid": {
            "type": "string",
            "example": "s-success-rid"
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "meta": {
            "type": "object",
            "example": {
              "timestamp": "2026-03-28T07:44:00Z"
            }
          }
        },
        "required": [
          "rid",
          "statusCode",
          "data",
          "message",
          "meta"
        ]
      },
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Villa Fit-out - Sector 45",
            "description": "The title of the project"
          },
          "description": {
            "type": "string",
            "example": "Full interior fit-out including MEP and furniture"
          },
          "status": {
            "type": "string",
            "example": "active",
            "description": "Initial status of the project"
          },
          "estimatedValue": {
            "type": "number",
            "example": 1500000,
            "description": "Budget/Estimated value"
          },
          "budget": {
            "type": "number",
            "example": 1500000,
            "description": "Alias for estimated value used by the frontend contract"
          },
          "contactId": {
            "type": "string",
            "example": "uuid-contact"
          },
          "leadId": {
            "type": "string",
            "example": "uuid-lead"
          },
          "projectManagerId": {
            "type": "string",
            "example": "uuid",
            "description": "Assigned Project Manager"
          },
          "managerId": {
            "type": "string",
            "example": "uuid",
            "description": "Alias for assigned project manager"
          },
          "startDate": {
            "type": "string",
            "example": "2026-04-01T00:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "example": "2026-08-01T00:00:00.000Z"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Revised Villa Fit-out"
          },
          "description": {
            "type": "string",
            "example": "Revised scope with added joinery package"
          },
          "status": {
            "type": "string",
            "example": "completed"
          },
          "estimatedValue": {
            "type": "number",
            "example": 1750000
          },
          "budget": {
            "type": "number",
            "example": 1750000
          },
          "contactId": {
            "type": "string",
            "example": "uuid-contact"
          },
          "leadId": {
            "type": "string",
            "example": "uuid-lead"
          },
          "projectManagerId": {
            "type": "string",
            "example": "uuid-user"
          },
          "managerId": {
            "type": "string",
            "example": "uuid-user"
          },
          "startDate": {
            "type": "string",
            "example": "2026-04-01T00:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "example": "2026-08-01T00:00:00.000Z"
          }
        }
      },
...
      "UpdateRolePermissionsDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}